Demo Day 01-06-22

+++ burger-queen/src/components/Admin/Crud/Read.jsx
@@ -7,7 +7,7 @@ import deleteIcon from "../../../img/OrderStatus/deleteOrder.svg";
 import { useState, useEffect } from "react";
 import CreateNewUser from "./CreateNewUser";
 import CreateNewProduct from "./CreateNewProduct";
-import Delete from "./Delete";
+
 
 export default function Read(prop) {
   let url = "";
@@ -40,6 +40,10 @@ export default function Read(prop) {
   }, [url]);
 
   function deleteUsers(users) {
+    let isDelete = window.confirm(
+      `¿Estás seguro de eliminar el registro con el usuario '${users.name}'?`
+    );
+    if (isDelete) {
     const requestOptions = {
       method: "DELETE",
       headers: { "Content-Type": "application/json" },
@@ -50,10 +54,19 @@ export default function Read(prop) {
     )
       .then((response) => response.json())
       .then((data) => console.log(data.name));
+      let newTable = listitem.filter((db) => db.name !== users.name);
+      setListItem(newTable);
+    }else {
+        return;
+      }
   }
 
   function deleteProduct(products) {
-    const requestOptions = {
+    let isDelete = window.confirm(
+      `¿Estás seguro de eliminar el registro con el producto '${products.item}'?`
+    );
+    if (isDelete) {
+       const requestOptions = {
       method: "DELETE",
       headers: { "Content-Type": "application/json" },
     };
@@ -63,6 +76,9 @@ export default function Read(prop) {
     )
       .then((response) => response.json())
       .then((data) => console.log(data.name));
+    }else {
+      return;
+    }
   }
 
   return (
@@ -72,33 +88,33 @@ export default function Read(prop) {
         url === "https://6290ec0e27f4ba1c65c4cd21.mockapi.io/api/user" ? (
           <section className="container-show-control-panel">
             <h3 className="titles-control-panel">Users List</h3>
-            <table >
+            <table className="data-control-panel" >
               <thead>
                 <tr>
-                  <th>Name</th>
-                  <th>Email</th>
-                  <th>Password</th>
-                  <th>Rol</th>
-                  <th>Edit</th>
-                  <th>Delete</th>
+                  <th className="column-control-panel">Name</th>
+                  <th className="column-control-panel">Email</th>
+                  <th className="column-control-panel">Password</th>
+                  <th className="column-control-panel">Rol</th>
+                  <th className="column-control-panel">Edit</th>
+                  <th className="column-control-panel">Delete</th>
                 </tr>
               </thead>
 
               {listitem.map((users) => (
                 <tr>
-                  <td>{users.name}</td>
-                  <td>{users.email}</td>
-                  <td type="password">{users.password}</td>
-                  <td>{users.rol}</td>
-                  <td>
+                  <td className="row-control-panel">{users.name}</td>
+                  <td className="row-control-panel">{users.email}</td>
+                  <td className="row-control-panel">{users.password}</td>
+                  <td className="row-control-panel">{users.rol}</td>
+                  <td className="row-control-panel">
                     <img
                       className="icon-edit"
                       src={editIcon}
                       alt="edit-icon"
-                      onClick={<Delete userid={users} />}
+                      onClick={handleClick}
                     />
                   </td>
-                  <td>
+                  <td className="row-control-panel">
                     <img
                       className="icon-delete"
                       src={deleteIcon}
@@ -109,7 +125,7 @@ export default function Read(prop) {
                   </td>
                 </tr>
               ))}
-
+              </table>
               <button
                 type="button"
                 className="btn-add-item"
@@ -118,38 +134,38 @@ export default function Read(prop) {
               >
                 Add New User
               </button>
-            </table>
+           
           </section>
         ) : (
           <section className="container-show-control-panel">
             <h3 className="titles-control-panel">Product List</h3>
-            <table>
+            <table  className="data-control-panel">
               <thead>
                 <tr>
-                  <th>Id</th>
-                  <th>Product</th>
-                  <th>Price</th>
-                  <th>Image</th>
-                  <th>Type</th>
-                  <th>Edit</th>
-                  <th>Delete</th>
+                  <th className="column-control-panel">Id</th>
+                  <th className="column-control-panel">Product</th>
+                  <th className="column-control-panel" >Price</th>
+                  <th className="column-control-panel" >Image</th>
+                  <th className="column-control-panel">Type</th>
+                  <th className="column-control-panel" >Edit</th>
+                  <th className="column-control-panel" >Delete</th>
                 </tr>
               </thead>
 
               {listitem.map((products) => (
                 <tr>
-                  <td>{products.id}</td>
-                  <td>{products.item}</td>
-                  <td>${products.price}</td>
-                  <td>
+                  <td className="row-control-panel">{products.id}</td>
+                  <td className="row-control-panel">{products.item}</td>
+                  <td className="row-control-panel">${products.price}</td>
+                  <td className="row-control-panel">
                     <img
                       className="icon-product"
                       src={products.img}
                       alt="-icon"
                     />
                   </td>
-                  <td>{products.type}</td>
-                  <td>
+                  <td className="row-control-panel">{products.type}</td>
+                  <td className="row-control-panel">
                     <img
                       className="icon-edit"
                       src={editIcon}
@@ -157,7 +173,7 @@ export default function Read(prop) {
                       onClick={handleClick}
                     />
                   </td>
-                  <td>
+                  <td className="row-control-panel" >
                     <img
                       className="icon-delete"
                       src={deleteIcon}
@@ -167,7 +183,7 @@ export default function Read(prop) {
                   </td>
                 </tr>
               ))}
-
+           </table>
               <button
                 type="button"
                 className="btn-add-item"
@@ -176,7 +192,7 @@ export default function Read(prop) {
               >
                 Add New Product
               </button>
-            </table>
+           
           </section>
         )
       ) : (
